linux编译错误

+++ network/async_common_darwin.go
@@ -4,7 +4,7 @@ package network
 
 // macOS平台的事件常量（兼容性定义）
 const (
-	EpollIn  = 1  // 兼容EPOLLIN
-	EpollOut = 4  // 兼容EPOLLOUT  
-	EpollET  = 0  // macOS kqueue本身就是边缘触发
+	EpollIn  = uint32(1)  // 兼容EPOLLIN
+	EpollOut = uint32(4)  // 兼容EPOLLOUT  
+	EpollET  = uint32(0)  // macOS kqueue本身就是边缘触发
 )
